// frontend/src/app/core/services/api/model/clubCoachDto.ts
/**
 * BFG API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ClubDto } from './clubDto';
import { UserDto } from './userDto';


/**
 * Data Transfer Object representing a coach assignment to a club
 */
export interface ClubCoachDto { 
    /**
     * Unique identifier of the coach assignment
     */
    uuid?: string;
    /**
     * UUID of the user (coach)
     */
    userId?: string;
    /**
     * UUID of the club
     */
    clubId?: string;
    /**
     * Timestamp when the coach was assigned to the club
     */
    assignmentDate?: string;
    coach?: UserDto;
    club?: ClubDto;
}